fixed macros to not assume floating type

+++ tests/trace.rs
@@ -5,11 +5,11 @@ trace_macros!(true);
 
 #[test]
 fn form_basic() {
-    let x: [f64; 4] = [1.0, 2.0, 3.0, 4.0];
-    let y: [f64; 4] = [4.0, 3.0, 2.0, 1.0];
-    let mut z: [f64; 4] = [0., 0., 0., 0.];
+    let x: [i32; 4] = [1, 2, 3, 4];
+    let y: [i32; 4] = [4, 3, 2, 1];
+    let mut z: [i32; 4] = [0, 0, 0, 0];
     axpy![z = x + y];
-    assert_eq!(z, [5f64, 5., 5., 5.]);
+    assert_eq!(z, [5, 5, 5, 5]);
 }
 
 #[test]
